+++ pkg/session/BUILD.bazel
@@ -105,7 +105,6 @@ go_library(
         "//pkg/tablecodec",
         "//pkg/telemetry",
         "//pkg/testkit/testenv",
-        "//pkg/testkit/testutil",
         "//pkg/types",
         "//pkg/types/parser_driver",
         "//pkg/util",
