+++ pkg/testkit/BUILD.bazel
@@ -42,7 +42,6 @@ go_library(
         "//pkg/store/helper",
         "//pkg/store/mockstore",
         "//pkg/testkit/testenv",
-        "//pkg/testkit/testutil",
         "//pkg/types",
         "//pkg/util",
         "//pkg/util/breakpoint",
