+++ pkg/testkit/testenv/BUILD.bazel
@@ -5,4 +5,8 @@ go_library(
     srcs = ["testenv.go"],
     importpath = "github.com/pingcap/tidb/pkg/testkit/testenv",
     visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/config",
+        "//pkg/keyspace",
+    ],
 )
