+++ pkg/testkit/testenv/testenv.go
@@ -16,9 +16,28 @@ package testenv
 
 import (
 	"runtime"
+	"testing"
+
+	"github.com/pingcap/tidb/pkg/config"
+	"github.com/pingcap/tidb/pkg/keyspace"
 )
 
 // SetGOMAXPROCSForTest sets GOMAXPROCS to 16 if it is greater than 16.
 func SetGOMAXPROCSForTest() {
 	runtime.GOMAXPROCS(min(16, runtime.GOMAXPROCS(0)))
+}
+
+// UpdateConfigForNextgen updates the global config to use SYSTEM keyspace.
+func UpdateConfigForNextgen(t testing.TB) {
+	t.Helper()
+	// in nextgen, SYSTEM ks must be bootstrapped first, to make UT easier, we
+	// always run them inside SYSTEM keyspace, if your test requires bootstrapping
+	// multiple keyspace, you should use CreateMockStoreAndDomainForKS instead.
+	bak := *config.GetGlobalConfig()
+	t.Cleanup(func() {
+		config.StoreGlobalConfig(&bak)
+	})
+	config.UpdateGlobal(func(conf *config.Config) {
+		conf.KeyspaceName = keyspace.System
+	})
 }
