+++ pkg/testkit/testutil/BUILD.bazel
@@ -10,8 +10,6 @@ go_library(
     importpath = "github.com/pingcap/tidb/pkg/testkit/testutil",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/config",
-        "//pkg/keyspace",
         "//pkg/kv",
         "//pkg/parser/mysql",
         "//pkg/sessionctx/stmtctx",
